backend/requirements.txt

fastapi
uvicorn
spacy
openai
python-dotenv
